requirements.txt
flask == 2.3.3
flask-mysqldb
pip == 1.0.4
PyJWT = 2.8.0

flask-cors